src: M1 rollout point 9: -pi typed dispatch for atk/basics/common; rock retypes + tree-wide consumer casts
FossilOrigin-Name: 3a974e5efb0f2f5e928b92d2fee4ca96202accde050147ea267deb0e5113a2ce

+++ src/atk/adew/pagev.c
@@ -318,7 +318,7 @@ struct menulist *ml;
 	    menulist_DeleteFromML(self->ml, MenuBuf);
 	} else {
 	    menulist_AddToML(self->ml, MenuBuf,
-			     switchobjproc, (long) sw, 0);
+			     switchobjproc, sw, 0);
 	}
     }
     if (ml) menulist_ChainBeforeML(self->ml, ml, ml);
